refactor: domain.check moved from domain
