+++ src/main/kotlin/domain/WinningLotto.kt
@@ -1,5 +1,7 @@
 package org.example.domain
 
+import org.example.domain.check.BonusLottoNumbersCheck
+
 class WinningLotto(
     numbers: List<Int>,
     val bonusNumber: Int,
