+++ src/main/kotlin/domain/check/BonusLottoNumbersCheck.kt
@@ -1,4 +1,4 @@
-package org.example.domain
+package org.example.domain.check
 
 import org.example.config.LottoConfig
